% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxa_and_mbness.R
\name{check_taxa_and_mbness}
\alias{check_taxa_and_mbness}
\title{Check results}
\usage{
check_taxa_and_mbness(crown_age = 8)
}
\value{
Check that results will match conditions
}
\description{
Check results
}
\author{
Giovanni Laudanno
}
