% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(bins, cond, crown_age, lambda, lambdas, mu, mus,
  n_replicates, nu, nus, q, qs, saveit, seed, var, verbose)
}
\arguments{
\item{bins}{number of bins}

\item{cond}{Conditioning}

\item{crown_age}{The crown age of the tree}

\item{lambda}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{lambdas}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mu}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{mus}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{n_replicates}{number of replicates}

\item{nu}{the rate at which a multiple-birth specation is triggered.}

\item{nus}{the rate at which a multiple-birth specation is triggered.}

\item{q}{per-species speciation probability in case of occurrance of}

\item{qs}{per-species speciation probability in case of occurrance of}

\item{saveit}{do you want to save it?}

\item{seed}{a random number generator seed}

\item{var}{a variable}

\item{verbose}{give more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno.
}
