% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_taxa_and_mbness.R
\name{measure_taxa_and_mbness}
\alias{measure_taxa_and_mbness}
\title{Measure number of taxa and mb-ness for each parameter setting}
\usage{
measure_taxa_and_mbness(lambdas = c(0.2), mus = c(0, 0.1), nus = c(0,
  0.5, 1, 1.5), qs = c(0.1, 0.15, 0.2), cond = 1, crown_age = 8,
  n_replicates = 10000, saveit = TRUE)
}
\value{
a dataframe with parameter settings, a sample of number of taxa and
 a sample of mb-ness (in percentage and absolute).
}
\description{
Measure number of taxa and mb-ness for each parameter setting
}
\author{
Giovanni Laudanno
}
