% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_and_mbness.R
\name{plot_taxa_and_mbness}
\alias{plot_taxa_and_mbness}
\title{Plot number of taxa and mb-ness for each parameter setting}
\usage{
plot_taxa_and_mbness(lambdas = c(0.2), mus = c(0, 0.1), nus = c(0,
  0.5, 1, 1.5), qs = c(0.1, 0.15, 0.2), cond = 1, crown_age = 8,
  bins = 15, saveit = TRUE)
}
\value{
plots for all the variables
}
\description{
Plot number of taxa and mb-ness for each parameter setting
}
\author{
Giovanni Laudanno
}
