% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_taxa_and_mbness.R
\name{analyze_taxa_and_mbness}
\alias{analyze_taxa_and_mbness}
\title{Analyze results}
\usage{
analyze_taxa_and_mbness(crown_age = 8)
}
\value{
List with meaningful measures for each parameter setting
}
\description{
Analyze results
}
\author{
Giovanni Laudanno
}
