% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_taxa_and_mbness.R
\name{analyze_taxa_and_mbness}
\alias{analyze_taxa_and_mbness}
\title{Analyze results}
\usage{
analyze_taxa_and_mbness(lambdas = c(0.2), mus = c(0, 0.1), nus = c(0,
  0.5, 1, 1.5), qs = c(0.1, 0.15, 0.2), cond = 1, crown_age = 8)
}
\arguments{
\item{lambdas}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mus}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{nus}{the rate at which a multiple-birth specation is triggered.}

\item{qs}{per-species speciation probability in case of occurrance of}

\item{cond}{Conditioning}

\item{crown_age}{The crown age of the tree}
}
\value{
List with meaningful measures for each parameter setting
}
\description{
Analyze results
}
\author{
Giovanni Laudanno
}
