% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxa_and_mbness.R
\name{check_taxa_and_mbness}
\alias{check_taxa_and_mbness}
\title{Check results}
\usage{
check_taxa_and_mbness(lambdas = c(0.2), mus = c(0, 0.1), nus = c(0,
  0.5, 1, 1.5), qs = c(0.1, 0.15, 0.2), cond = 1, crown_age = 8)
}
\value{
Check that results will match conditions
}
\description{
Check results
}
\author{
Giovanni Laudanno
}
