% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_taxa_and_mbness.R
\name{plot_taxa_and_mbness}
\alias{plot_taxa_and_mbness}
\title{Plot number of taxa and mb-ness for each parameter setting}
\usage{
plot_taxa_and_mbness()
}
\value{
plots for all the variables
}
\description{
Plot number of taxa and mb-ness for each parameter setting
}
\author{
Giovanni Laudanno
}
